pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isage-sias"
version = "0.1.0"
description = "Sample-Importance-Aware Selection (SIAS) - Continual learning and coreset selection algorithms"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "IntelliStream Team", email = "shuhao_zhang@hust.edu.cn" }
]
keywords = [
    "sias",
    "sample-selection",
    "importance-aware",
    "continual-learning",
    "coreset",
    "active-learning",
    "data-selection",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "numpy>=1.20.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.4",
    "mypy>=1.0.0",
]

torch = [
    "torch>=2.0.0",
]

all = [
    "isage-sias[dev,torch]",
]

[project.urls]
Homepage = "https://github.com/intellistream/sage-sias"
Repository = "https://github.com/intellistream/sage-sias"
Documentation = "https://github.com/intellistream/sage-sias#readme"
"Bug Tracker" = "https://github.com/intellistream/sage-sias/issues"

[tool.setuptools]
packages = ["sage_sias"]

[tool.setuptools.package-dir]
sage_sias = "."

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
